+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/OpenAs.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.overdue.download.formatdialog
+
+sealed class OpenAs
+
+object Share : OpenAs()
+
+object Download : OpenAs()