+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/SelectOverdueDownloadFormatEvent.kt
@@ -1,0 +1,13 @@
+package org.simple.clinic.overdue.download.formatdialog
+
+import android.net.Uri
+
+sealed class SelectOverdueDownloadFormatEvent
+
+data class FileDownloadedForSharing(val uri: Uri) : SelectOverdueDownloadFormatEvent()
+
+object DownloadOrShareClicked : SelectOverdueDownloadFormatEvent()
+
+object OverdueDownloadScheduled : SelectOverdueDownloadFormatEvent()
+
+object CancelClicked : SelectOverdueDownloadFormatEvent()