+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/UiActions.kt
@@ -1,0 +1,8 @@
+package org.simple.clinic.overdue.download.formatdialog
+
+import android.net.Uri
+
+interface UiActions {
+  fun shareDownloadedFile(downloadedUri: Uri)
+  fun dismiss()
+}