+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Changes
 
-- [In Progress: 12 Nov 2021] Add option to download & share overdue list
+- [In Progress: 18 Nov 2021] Add option to download & share overdue list
 
 ## On Demo